name consolidation
This project refers to itself, variously, as 'bmap-tools', 'bmap-tool',
'bmaptools', 'Bmaptool', etc. Going forward this project/tool/etc is to be
known, singularly, as 'bmaptool'.
