+++ bmaptool/__main__.py
@@ -2,7 +2,7 @@
 import re
 import sys
 
-from bmaptools.CLI import main
+from .CLI import main
 
 if __name__ == "__main__":
     sys.argv[0] = re.sub(r"(-script\.pyw|\.exe|\.pyz)?$", "", sys.argv[0])
