+++ debian/bmaptool.docs
