+++ debian/bmaptool.install
