+++ debian/copyright
@@ -1,5 +1,5 @@
 Format: http://www.debian.org/doc/packaging-manuals/copyright-format/1.0/
-Upstream-name: bmap-tools
+Upstream-name: bmaptool
 Upstream-Contact: Trevor Woerner <twoerner@gmail.com>
 Source: https://github.com/yoctoproject/bmaptool
   .
