+++ docs/man1/bmaptool.1
@@ -1,4 +1,4 @@
-.TH BMAPTOOL "1" "February 2014" "bmap-tools 3.2" "User Commands"
+.TH BMAPTOOL "1" "February 2014" "bmaptool 3.2" "User Commands"
 
 .SH NAME
 
@@ -14,13 +14,13 @@ bmaptool - create block map (bmap) for a file or copy a file using bmap
 .SH DESCRIPTION
 
 .PP
-\fIBmaptool\fR is a generic tool for creating the block map (bmap) for a file and
+\fIbmaptool\fR is a generic tool for creating the block map (bmap) for a file and
 copying files using the block map. The idea is that large files, like raw
 system image files, can be copied or flashed a lot faster with \fIbmaptool\fR than
 with traditional tools, like "dd" or "cp".
 
 .PP
-\fIBmaptool\fR supports 2 commands:
+\fIbmaptool\fR supports 2 commands:
 .RS 2
 1. \fBcopy\fR - copy a file to another file using bmap or flash an image to a block device
 .RE
