+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "bmaptools"
+name = "bmaptool"
 version = "3.7.0"
 license = "GPL-2.0-only"
 description = "BMAP tools"
@@ -15,7 +15,7 @@ black = "^22.3.0"
 nose = "^1.3.7"
 
 [tool.poetry.scripts]
-bmaptool = "bmaptools.CLI:main"
+bmaptool = "bmaptool.CLI:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
