+++ setup.cfg
@@ -1,2 +1,2 @@
 [nosetests]
-cover-package=bmaptools
+cover-package=bmaptool
