+++ setup.py
@@ -6,7 +6,7 @@ from setuptools import setup, find_packages
 
 def get_version():
     """Fetch the project version number from the 'bmaptool' file."""
-    with open("bmaptools/CLI.py", "r") as fobj:
+    with open("bmaptool/CLI.py", "r") as fobj:
         for line in fobj:
             matchobj = re.match(r'^VERSION = "(\d+.\d+)"$', line)
             if matchobj:
@@ -16,7 +16,7 @@ def get_version():
 
 
 setup(
-    name="bmap-tools",
+    name="bmaptool",
     description="Tools to generate block map (AKA bmap) and copy images " "using bmap",
     author="Artem Bityutskiy",
     author_email="artem.bityutskiy@linux.intel.com",
@@ -24,12 +24,12 @@ setup(
     maintainer_email="twoerner@gmail.com",
     version=get_version(),
     entry_points={
-        "console_scripts": ["bmaptool=bmaptools.CLI:main"],
+        "console_scripts": ["bmaptool=bmaptool.CLI:main"],
     },
     packages=find_packages(exclude=["test*"]),
     license="GPLv2",
     long_description="Tools to generate block map (AKA bmap) and flash "
-    "images using bmap. Bmaptool is a generic tool for "
+    "images using bmap. bmaptool is a generic tool for "
     "creating the block map (bmap) for a file, and copying "
     "files using the block map. The idea is that large file "
     "containing unused blocks, like raw system image files, "
